requirements.txt
fastapi
uvicorn
openai-whisper
torch
python-docx
reportlab
python-multipart